reafctor: rename files updateStudents to updateAdmin
